feat: comment 예외 로직 적용

+++ src/main/java/com/skhuedin/skhuedin/controller/api/CommentApiController.java
@@ -4,10 +4,12 @@ import com.skhuedin.skhuedin.controller.response.BasicResponse;
 import com.skhuedin.skhuedin.controller.response.CommonResponse;
 import com.skhuedin.skhuedin.dto.comment.CommentMainResponseDto;
 import com.skhuedin.skhuedin.dto.comment.CommentSaveRequestDto;
+import com.skhuedin.skhuedin.security.AuthService;
 import com.skhuedin.skhuedin.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -27,10 +29,16 @@ import java.util.List;
 public class CommentApiController {
 
     private final CommentService commentService;
+    private final AuthService authService;
 
     @PostMapping("comments")
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> save(@Valid @RequestBody CommentSaveRequestDto requestDto) {
+
+        if (!authService.isSameUser(requestDto.getWriterUserId())) {
+            throw new AccessDeniedException("동일하지 않은 유저 정보입니다.");
+        }
+
         Long saveId = commentService.save(requestDto);
 
         return ResponseEntity.status(HttpStatus.CREATED)
@@ -48,6 +56,11 @@ public class CommentApiController {
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> update(@PathVariable("commentId") Long id,
                                                           @Valid @RequestBody CommentSaveRequestDto updateDto) {
+
+        if (!authService.isSameUser(updateDto.getWriterUserId())) {
+            throw new AccessDeniedException("동일하지 않은 유저 정보입니다.");
+        }
+
         Long commentId = commentService.update(id, updateDto);
         CommentMainResponseDto responseDto = commentService.findById(commentId);
 
@@ -57,6 +70,13 @@ public class CommentApiController {
     @DeleteMapping("comments/{commentId}")
     @PreAuthorize("hasAnyRole('ROLE_USER','ROLE_ADMIN')")
     public ResponseEntity<? extends BasicResponse> delete(@PathVariable("commentId") Long id) {
+
+        CommentMainResponseDto commentMainResponseDto = commentService.findById(id);
+
+        if (!authService.isSameUser(commentMainResponseDto.getWriterUser().getId())) {
+            throw new AccessDeniedException("동일하지 않은 유저 정보입니다.");
+        }
+
         commentService.delete(id);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
